fix(terminal): remove discovery legend above hybrid input
- Remove showLegend derivation and legend DOM element
- Narrow absolute container gate to isFleetPrimary only
- Rename test file, preserving placeholder tests, dropping legend tests
